requirements.txt
openai
python-dotenv
requests
poml